requirements.txt
accelerate @ git+https://github.com/huggingface/accelerate.git@1b50d93999efdfb47eeae71395c921ef12c85d02
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
datasets==4.1.0
Deprecated==1.2.18
dill==0.4.0
filelock==3.13.1
fonttools==4.60.0
frozenlist==1.7.0
fsspec==2024.6.1
granite-tsfm==0.3.1
hf-xet==1.1.10
huggingface-hub==0.35.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.4
joblib==1.5.2
kiwisolver==1.4.9
llvmlite==0.44.0
MarkupSafe==2.1.5
matplotlib==3.10.6
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
networkx==3.3
numba==0.61.2
numpy==2.1.2
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.4.1.4
nvidia-cufile-cu12==1.14.1.1
nvidia-curand-cu12==10.3.10.19
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.9.79
packaging==25.0
pandas==2.3.2
pillow==11.0.0
pluggy==1.6.0
propcache==0.3.2
psutil==7.0.0
pyarrow==21.0.0
Pygments==2.19.2
pynndescent==0.5.13
pyparsing==3.2.4
pytest==8.4.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
setuptools==70.2.0
six==1.17.0
sympy==1.13.3
threadpoolctl==3.6.0
tokenizers==0.22.0
torch==2.8.0+cu129
torchvision==0.23.0+cu129
tqdm==4.67.1
transformers==4.56.1
triton==3.4.0
typing_extensions==4.12.2
tzdata==2025.2
umap-learn==0.5.9.post2
urllib3==2.5.0
wrapt==1.17.3
xxhash==3.5.0
yarl==1.20.1
